Update BusChangeMachineClass.h
오류해결

+++ Bus Change Machine Class/Bus Change Machine Class/BusChangeMachineClass.h
@@ -179,11 +179,11 @@ public:
 
 private:
 
-	int adult;
+	float adult;
 
-	int teenager;
+	float teenager;
 
-	int child;
+	float child;
 
 	int money;
 
@@ -191,7 +191,7 @@ private:
 
 	int age;
 
-	int discount = 0;
+	float discount = 0;
 
 	bool moneycheck;
